requirements.txt
pillow==9.5.0
pandas==1.5.3